+++ Code/src/network/transport/ipc.c
@@ -1,35 +1,35 @@
 #include "ipc.h"
 
-static int writeAll(int fd, const void *buffer, size_t size) {
+static int writeAll(int fd, const void *buffer, unsigned long size) {
     const uint8_t *cursor = (const uint8_t *)buffer;
-    size_t totalWritten = 0;
+    unsigned long totalWritten = 0;
 
     while (totalWritten < size) {
-        ssize_t written = write(fd, cursor + totalWritten, size - totalWritten);
+        long written = write(fd, cursor + totalWritten, size - totalWritten);
         if (written <= 0) {
             return -1;
         }
-        totalWritten += (size_t)written;
+        totalWritten += (unsigned long)written;
     }
 
     return 0;
 }
 
 // Lectura bloquejant completa: el canal IPC Maester<->Envoy és un pipe en el qual
-// l'envoy espera ociós una missió (potser molt de temps). NO posem timeout aquí:
+// l'envoy espera  una missió (potser molt de temps). NO posem timeout aquí:
 // si poséssim, l'envoy "moriria" tot sol després d'uns segons sense feina i un
 // write() posterior del Maester rebria SIGPIPE. read() retorna 0 (EOF) si l'altre
 // extrem tanca el pipe (p. ex. el Maester acaba), cosa que sí trenca el bucle net.
-static int readAll(int fd, void *buffer, size_t size) {
+static int readAll(int fd, void *buffer, unsigned long size) {
     uint8_t *cursor = (uint8_t *)buffer;
-    size_t totalRead = 0;
+    unsigned long totalRead = 0;
 
     while (totalRead < size) {
-        ssize_t bytesRead = read(fd, cursor + totalRead, size - totalRead);
+        unsigned long bytesRead = read(fd, cursor + totalRead, size - totalRead);
         if (bytesRead <= 0) {
             return -1;
         }
-        totalRead += (size_t)bytesRead;
+        totalRead += (unsigned long)bytesRead;
     }
 
     return 0;
